diabetes/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
joblib==1.3.2
numpy==1.26.3
Pillow==9.3.0
psycopg2-binary==2.9.6
PyJWT==2.1.0
pytz==2021.1
scikit-learn==1.4.0
scipy==1.12.0
sqlparse==0.4.1
sudo==1.0.0
threadpoolctl==3.2.0
tzdata==2022.6
whitenoise==5.3.0
